+++ test/services/offenderService.test.js
@@ -3496,43 +3496,6 @@ describe('isInitialInProgress', () => {
   })
 })
 
-describe('isOverdue', () => {
-  it('returns true if the date is before the current time', () => {
-    const dbDate = '2022-01-01'
-    const now = moment('2022-01-02', 'YYYY-MM-DD')
-    jest.spyOn(moment, 'now').mockImplementation(() => now.valueOf())
-
-    expect(service.isOverdue(dbDate)).toBe(true)
-  })
-
-  it('returns false if the date is after the current time', () => {
-    const dbDate = '2022-01-02'
-    const now = moment('2022-01-01', 'YYYY-MM-DD')
-    jest.spyOn(moment, 'now').mockImplementation(() => now.valueOf())
-
-    expect(service.isOverdue(dbDate)).toBe(false)
-  })
-
-  it('returns false if the date is the same as the current time', () => {
-    const dbDate = '2022-01-01'
-    const now = moment('2022-01-01', 'YYYY-MM-DD')
-    jest.spyOn(moment, 'now').mockImplementation(() => now.valueOf())
-
-    expect(service.isOverdue(dbDate)).toBe(false)
-  })
-
-  it('returns false if the input is null or undefined', () => {
-    expect(service.isOverdue(null)).toBe(false)
-    expect(service.isOverdue(undefined)).toBe(false)
-  })
-
-  it('returns false if the input is not a string in the format of "YYYY-MM-DD"', () => {
-    expect(service.isOverdue('2022-01-01T00:00:00')).toBe(false)
-    expect(service.isOverdue('January 1, 2022')).toBe(false)
-    expect(service.isOverdue('22-01-01')).toBe(false)
-  })
-})
-
 describe('calculateRecatDisplayStatus', () => {
   it('returns "Not started" when given an empty or undefined displayStatus', () => {
     expect(service.calculateRecatDisplayStatus(undefined)).toBe('Not started')
